current millis is now exactly before future!

+++ src/main/java/ir/hajk1/n26/challenge/service/TransactionService.java
@@ -1,10 +1,9 @@
 package ir.hajk1.n26.challenge.service;
 
 import ir.hajk1.n26.challenge.model.Transaction;
-
-import javax.validation.Valid;
 import java.util.List;
 import java.util.concurrent.ConcurrentMap;
+import javax.validation.Valid;
 
 /**
  * Created by k1 on 6/29/18.
@@ -13,6 +12,6 @@ import java.util.concurrent.ConcurrentMap;
 public interface TransactionService {
     void persist(@Valid Transaction transaction);
 
-    ConcurrentMap<Long, List<Double>> getTransactionMap();
+    ConcurrentMap<Byte, List<Double>> getTransactionMap();
 
 }
