+++ src/main/java/ir/hajk1/n26/challenge/service/TransactionServiceImpl.java
@@ -1,12 +1,6 @@
 package ir.hajk1.n26.challenge.service;
 
 import ir.hajk1.n26.challenge.model.Transaction;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.scheduling.annotation.Scheduled;
-import org.springframework.stereotype.Service;
-
-import javax.validation.Valid;
 import java.time.Instant;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
@@ -15,6 +9,11 @@ import java.util.List;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 import java.util.function.BiConsumer;
+import javax.validation.Valid;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.scheduling.annotation.Scheduled;
+import org.springframework.stereotype.Service;
 
 /**
  * Created by k1 on 6/29/18.
@@ -24,22 +23,22 @@ import java.util.function.BiConsumer;
 public class TransactionServiceImpl implements TransactionService {
     private static final Logger logger = LoggerFactory.getLogger(TransactionServiceImpl.class);
 
-    private ConcurrentMap<Long, List<Double>> transactionMap;
+    private ConcurrentMap<Byte, List<Double>> transactionMap;
 
     public TransactionServiceImpl() {
         transactionMap = new ConcurrentHashMap<>(60);
     }
 
-    public ConcurrentMap<Long, List<Double>> getTransactionMap() {
+    public ConcurrentMap<Byte, List<Double>> getTransactionMap() {
         return transactionMap;
     }
 
     private void addTransaction(Transaction transaction) {
         byte second = extractSecondPortion(transaction.getTimestamp());
-        final BiConsumer<? super Long, ? super List<Double>> action =
+        final BiConsumer<? super Byte, ? super List<Double>> action =
                 (key, value) -> transactionMap.computeIfAbsent(key, x -> new ArrayList<>())
                         .add(transaction.getAmount());
-        action.accept((long) second, new ArrayList<>());
+        action.accept(second, new ArrayList<>());
     }
 
     private byte extractSecondPortion(long timestamp) {
@@ -58,7 +57,7 @@ public class TransactionServiceImpl implements TransactionService {
     @Scheduled(cron = "0/1 * * * * *")
     public void cronJob() {
         byte currentSecond = (byte) LocalDateTime.now().getSecond();
-        transactionMap.put((long) currentSecond, new ArrayList<>());
+        transactionMap.put(currentSecond, new ArrayList<>());
         if (logger.isDebugEnabled())
             logger.debug("resetting Amounts for expired second index:" + currentSecond);
     }
